src/archivedb/requirements.txt
cbor==1.0.0
numexpr==2.7.0
numpy==1.17.4
Pillow==6.2.1
tables==3.6.1
